TLS methods refactor to allow create several Certificates
