+++ pkg/controller/cluster/minio.go
@@ -66,6 +66,7 @@ func (c *Controller) checkAndCreateMinIOCSR(ctx context.Context, nsName types.Na
 	return nil
 }
 
+// deleteCSR Removes a CSR
 func (c *Controller) deleteCSR(ctx context.Context, csrName string) error {
 	if certificates.GetCertificatesAPIVersion(c.kubeClientSet) == certificates.CSRV1 {
 		if err := c.kubeClientSet.CertificatesV1().CertificateSigningRequests().Delete(ctx, csrName, metav1.DeleteOptions{}); err != nil {
@@ -298,13 +299,7 @@ func (c *Controller) certNeedsRenewal(tlsSecret *corev1.Secret) (bool, error) {
 	var certPublicKey []byte
 	var certPrivateKey []byte
 
-	publicKey := "public.crt"
-	privateKey := "private.key"
-
-	if tlsSecret.Type == "kubernetes.io/tls" || tlsSecret.Type == "cert-manager.io/v1alpha2" || tlsSecret.Type == "cert-manager.io/v1" {
-		publicKey = "tls.crt"
-		privateKey = "tls.key"
-	}
+	publicKey, privateKey := c.getKeyNames(tlsSecret)
 
 	if _, exist := tlsSecret.Data[publicKey]; !exist {
 		return false, fmt.Errorf("missing '%s' in %s/%s secret", publicKey, tlsSecret.Namespace, tlsSecret.Name)
